bug end session
